# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/variants/{id}/stock-mode": {
      "patch": {
        "operationId": "AdminStockController_updateStockMode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStockModeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductVariantResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-stock"
        ]
      }
    },
    "/admin/variants/{id}/inventory": {
      "patch": {
        "operationId": "AdminStockController_updateInventory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInventoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-stock"
        ]
      }
    },
    "/admin/variants/{id}/add-stock": {
      "post": {
        "operationId": "AdminStockController_addStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "integer"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-stock"
        ]
      }
    },
    "/admin/inventory": {
      "get": {
        "operationId": "AdminStockController_findInventory",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stockMode",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-stock"
        ]
      }
    },
...
    "/admin/coupons": {
      "post": {
        "operationId": "AdminCouponController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCouponDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CouponResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-coupons"
        ]
      },
      "get": {
        "operationId": "AdminCouponController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CouponListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-coupons"
        ]
      }
    },
    "/admin/coupons/{id}": {
      "get": {
        "operationId": "AdminCouponController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CouponResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-coupons"
        ]
      },
      "patch": {
        "operationId": "AdminCouponController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCouponDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CouponResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-coupons"
        ]
      },
      "delete": {
        "operationId": "AdminCouponController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-coupons"
        ]
      }
    },
    "/admin/coupons/{id}/toggle-active": {
      "patch": {
        "operationId": "AdminCouponController_toggleActive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CouponResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-coupons"
        ]
      }
    },
...
  },
  "info": {
    "title": "E-commerce API",
    "description": "API para tienda y panel administrativo",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProductVariantResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sku": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "garmentType": {
            "type": "string"
          },
          "stockMode": {
            "type": "string"
          },
          "productionLeadTimeDays": {
            "type": "number"
          },
          "priceAdjustment": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sku",
          "stockMode",
          "productionLeadTimeDays",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateStockModeDto": {
        "type": "object",
        "properties": {
          "stockMode": {
            "type": "string",
            "enum": [
              "MADE_TO_ORDER",
              "TRACKED"
            ]
          }
        },
        "required": [
          "stockMode"
        ]
      },
      "UpdateInventoryDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "quantity"
        ]
      },
      "InventoryProductSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "InventorySummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "number"
          },
          "reservedQuantity": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "quantity",
          "reservedQuantity"
        ]
      },
      "InventoryListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sku": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "garmentType": {
            "type": "string"
          },
          "stockMode": {
            "type": "string"
          },
          "productionLeadTimeDays": {
            "type": "number"
          },
          "priceAdjustment": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "product": {
            "$ref": "#/components/schemas/InventoryProductSummaryDto"
          },
          "inventory": {
            "$ref": "#/components/schemas/InventorySummaryDto"
          }
        },
        "required": [
          "id",
          "sku",
          "stockMode",
          "productionLeadTimeDays",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "InventoryListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryListItemDto"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "limit": {
                "type": "number"
              },
              "total": {
                "type": "number"
              }
            },
            "required": [
              "page",
              "limit",
              "total"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "CreateCouponDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "discountType": {
            "type": "string",
            "enum": [
              "PERCENTAGE",
              "FIXED"
            ]
          },
          "discountValue": {
            "type": "number",
            "minimum": 0
          },
          "validFrom": {
            "type": "string"
          },
          "validUntil": {
            "type": "string"
          },
          "maxUses": {
            "type": "number",
            "minimum": 0
          },
          "isActive": {
            "type": "boolean"
          },
          "appliesTo": {
            "type": "string",
            "enum": [
              "ALL",
              "CATEGORY",
              "PRODUCT"
            ]
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "minOrderAmount": {
            "type": "number",
            "minimum": 0
          },
          "maxUsesPerUser": {
            "type": "number",
            "minimum": 0
          },
          "isFirstPurchaseOnly": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "discountType",
          "discountValue",
          "validFrom"
        ]
      },
      "CouponCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "CouponProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "CouponResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "discountType": {
            "type": "string"
          },
          "discountValue": {
            "type": "number"
          },
          "validFrom": {
            "type": "string"
          },
          "validUntil": {
            "type": "string"
          },
          "maxUses": {
            "type": "number"
          },
          "usedCount": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          },
          "appliesTo": {
            "type": "string"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "category": {
            "$ref": "#/components/schemas/CouponCategoryDto"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "product": {
            "$ref": "#/components/schemas/CouponProductDto"
          },
          "minOrderAmount": {
            "type": "number"
          },
          "maxUsesPerUser": {
            "type": "number"
          },
          "isFirstPurchaseOnly": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "discountType",
          "discountValue",
          "validFrom",
          "usedCount",
          "isActive",
          "appliesTo",
          "isFirstPurchaseOnly",
          "createdAt",
          "updatedAt"
        ]
      },
      "CouponListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CouponResponseDto"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "limit": {
                "type": "number"
              },
              "total": {
                "type": "number"
              }
            },
            "required": [
              "page",
              "limit",
              "total"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateCouponDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "discountType": {
            "type": "string",
            "enum": [
              "PERCENTAGE",
              "FIXED"
            ]
          },
          "discountValue": {
            "type": "number",
            "minimum": 0
          },
          "validFrom": {
            "type": "string"
          },
          "validUntil": {
            "type": "string",
            "nullable": true
          },
          "maxUses": {
            "type": "number",
            "minimum": 0
          },
          "isActive": {
            "type": "boolean"
          },
          "appliesTo": {
            "type": "string",
            "enum": [
              "ALL",
              "CATEGORY",
              "PRODUCT"
            ]
          },
          "categoryId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "minOrderAmount": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "maxUsesPerUser": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "isFirstPurchaseOnly": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}